refactor: JWT 쿠키세팅 로직 공통클래스로 분리
